+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Parameter `assay.name` for Seurat objects in `Cacoa$new` to allow selecting different assays
 - `cao$plotMetadataSeparation` function
 - `sample.subset` parameter for `plotSampleDistances` and `estimateMetadataSeparation`
+- `method="UMAP"` for `cao$plotSampleDistances`, which should work much better on larger sample collections (>50 or so)
 
 ### Changed